mcmc_skewnorm.py: Small changes to parameters.
mcmc_two_size.py: Added an overwrite flag and fixed the likelihood
error that is now 0.1 across all separations. Furthermore I started to
add a Poisson likelihood function but realized that it does not apply
in this case. Also messed around with the priors and jump sizes and
got a good fit with the two populations. Lastly, I've annotated the
trace plots.